requirements.txt
streamlit
google-cloud-aiplatform
google-generativeai
python-dotenv
cassandra-driver
langchain-google-genai
langchain